+++ app/Views/buyer/order.php
@@ -5,7 +5,7 @@
 <table>
     <tr>
         <th>Product</th>
-        <th>License</th>
+        <th>Purchased permissions</th>
         <th>Price</th>
         <th>Download</th>
     </tr>
@@ -17,7 +17,7 @@
            </a>
            </td>
            <td>
-           <?=H::e($i['license_type'])?>
+           <?=H::e($i['license_name'] ?: $i['license_type'])?><?php if(!empty($i['license_description'])):?><br><span class="muted"><?=H::e($i['license_description'])?></span><?php endif;?>
            </td>
            <td>
            <?=H::money($i['total_price'])?>